+++ secure-storage.android.d.ts
@@ -3,6 +3,9 @@ export declare class SecureStorage implements SecureStorageApi {
     private hawk;
     constructor();
     get(arg: GetOptions): Promise<any>;
+    getSync(arg: GetOptions): any;
     set(arg: SetOptions): Promise<boolean>;
+    setSync(arg: SetOptions): boolean;
     remove(arg: RemoveOptions): Promise<boolean>;
+    removeSync(arg: RemoveOptions): boolean;
 }