+++ secure-storage.android.ts
@@ -1,7 +1,7 @@
 import {SecureStorageApi, GetOptions, SetOptions, RemoveOptions} from "./secure-storage.common";
-const utils = require("utils/utils");
+const utils = require("tns-core-modules/utils/utils");
 
-declare var com: any;
+declare const com: any;
 
 export class SecureStorage implements SecureStorageApi {
 
@@ -12,10 +12,13 @@ export class SecureStorage implements SecureStorageApi {
   }
 
   public get(arg: GetOptions): Promise<any> {
-    let that = this;
     return new Promise((resolve, reject) => {
       resolve(com.orhanobut.hawk.Hawk.get(arg.key));
     });
+  }
+
+  getSync(arg: GetOptions): any {
+    return com.orhanobut.hawk.Hawk.get(arg.key);
   }
 
   public set(arg: SetOptions): Promise<boolean> {
@@ -24,9 +27,17 @@ export class SecureStorage implements SecureStorageApi {
     });
   }
 
+  setSync(arg: SetOptions): boolean {
+    return com.orhanobut.hawk.Hawk.put(arg.key, arg.value);
+  }
+
   public remove(arg: RemoveOptions): Promise<boolean> {
     return new Promise((resolve, reject) => {
       resolve(com.orhanobut.hawk.Hawk.delete(arg.key));
     });
+  }
+
+  removeSync(arg: RemoveOptions): boolean {
+    return com.orhanobut.hawk.Hawk.delete(arg.key);
   }
 }