+++ secure-storage.common.d.ts
@@ -13,6 +13,9 @@ export interface RemoveOptions {
 }
 export interface SecureStorageApi {
     get(arg: GetOptions): Promise<any>;
+    getSync(arg: GetOptions): any;
     set(arg: SetOptions): Promise<boolean>;
+    setSync(arg: SetOptions): boolean;
     remove(arg: RemoveOptions): Promise<boolean>;
+    removeSync(arg: RemoveOptions): boolean;
 }