+++ secure-storage.ios.d.ts
@@ -4,6 +4,9 @@ export declare class SecureStorage implements SecureStorageApi {
     private isSimulator;
     constructor();
     get(arg: GetOptions): Promise<any>;
+    getSync(arg: GetOptions): any;
     set(arg: SetOptions): Promise<boolean>;
+    setSync(arg: SetOptions): boolean;
     remove(arg: RemoveOptions): Promise<boolean>;
+    removeSync(arg: RemoveOptions): boolean;
 }